moved interceptor order tests
